requirements.txt
# claude-local-proxy requirements
# Install with: pip install -r requirements.txt

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Templates & static files
jinja2>=3.1.3
aiofiles>=23.2.1

# HTTP client (optional, stdlib urllib used as fallback)
httpx>=0.27.0

# Form parsing
python-multipart>=0.0.9

# Environment config
python-dotenv>=1.0.0

# Database (optional async driver, stdlib sqlite3 used as fallback)
aiosqlite>=0.20.0

# Minimal Termux install: pip install fastapi uvicorn jinja2 python-dotenv
